Base of PVP Fights
PVP Queue has been created. => NOW We have to make that the Fights are not locally but directly over the Server. Because now we can restart the fight over and over by pressing F5. We want to make the Server handle everything and the Users are just getting how it's going